Make setup-node faster and more reliable
We have occasionally seen that setup-node actions failing due to intermittent connection failures between us and nodejs.org.

To alleviate the issue, we could cache our "most-widely-used" Node.js version inside the container image, enabling setup-node to skip access to nodejs.org.

This tries to implements that, in a way so that it becomes resilient to nodejs and tool-cache changes. Just keep updating NODE_VERSION defined in the dockerfile using dependabot or renovate- the rest of the change will just work regardless of the version you want.